XXIX - bedExplosions gamerule. Optimised the code slightly. Fixed generic item group order. Added Cactus Flower to small_flowers, and Pitcher Plant to tall_flowers. Added Azalea wood (doesn't generate), ducks. Fixed zombie eyes not rendering properly. 7 new advancements: Passing the Time, Double Digits, MOAR Tools, Even MOAR Tools, This Isn't Punchout, Conservationist, Rainbow of Cuties. Added a couple items to a couple item groups

+++ src/main/java/net/epiccool/lawnchair/block/custom/WarpedWartBlock.java
@@ -48,12 +48,12 @@ public class WarpedWartBlock extends PlantBlock {
 
     @Override
     protected boolean hasRandomTicks(BlockState state) {
-        return (Integer)state.get(AGE) < 3;
+        return state.get(AGE) < 3;
     }
 
     @Override
     protected void randomTick(BlockState state, ServerWorld world, BlockPos pos, Random random) {
-        int i = (Integer)state.get(AGE);
+        int i = state.get(AGE);
         if (i < 3 && random.nextInt(10) == 0) {
             state = state.with(AGE, i + 1);
             world.setBlockState(pos, state, Block.NOTIFY_LISTENERS);
